+++ cppfd/boundary_conditions.h
@@ -4,7 +4,7 @@
 #include <cmath>
 #include <map>
 
-#include"meshchunk.h"
+#include"mesh_chunk.h"
 
 class BoundaryConditions
 {
